requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.1.1
Werkzeug==2.3.7
requests==2.26.0
python-dotenv==1.0.0
Pillow==10.0.0
SQLAlchemy==2.0.20
gunicorn==21.2.0 